+++ decidim-core/config/routes.rb
@@ -52,6 +52,8 @@ Decidim::Core::Engine.routes.draw do
 
   resources :pages, only: [:index, :show], format: false
 
+  get :organization_users, to: "users#index"
+
   get "/scopes/picker", to: "scopes#picker", as: :scopes_picker
 
   get "/static_map", to: "static_map#show", as: :static_map