+++ decidim-meetings/app/commands/decidim/meetings/admin/create_meeting.rb
@@ -43,6 +43,9 @@ module Decidim
             longitude: @form.longitude,
             location: @form.location,
             location_hints: @form.location_hints,
+            private_meeting: @form.private_meeting,
+            transparent: @form.transparent,
+            organizer: @form.organizer,
             registration_terms: @form.current_component.settings.default_registration_terms,
             component: @form.current_component
           )